Completes migration of plugin execution code in control to control/plugin
